+++ CHANGELOG.md
@@ -29,6 +29,8 @@ _Work toward next release. Currently under development._
 
 * Added support for syntax highlighting of inline code elements (#93 and #101)
 
+* Added the `ldap3` authentication type for login using an LDAPv3 server.
+
 **CHANGED:**
 
 * Auto-generated `csq_name` fields increment for every question, even those